fix: table style side effects
Spooky action at a distance was possible, thanks to each table style
serving as a global singleton. To resolve this, we return a copy every
time the user modifies the style.

A regression test is included. While it does not explicitly cover more
complex cases like altering the preset, asking for another, and then
asking for the original (which should be unaltered), it should suffice
to guard against this behavior.