Update application.py - Fixed running poetry --help
Running poetry --help was governo the behavior of poetry list --help instead of poetry list. 
The error was fixed in a simple way, changing the value of the variable "name" to "list" and removing the "help" parameter when calling the "_get_command_name(io)" funcition, when io assumed the values "-h" or "--help", in lines 370-371.
This is in response of Issue #8710.